// libs/routes/src/enums/route-experiment-key.ts
export enum RouteExperimentKey {
  CompanyRepresentativeProfile = 'staff_portal_company_representative_profile',
  CompanyRepresentativeCreate = 'staff_portal_create_company_representative',
  CompanyRepresentativeUpdate = 'staff_portal_update_company_representative',
  Talents = 'staff_portal_talent_list',
  TalentProfile = 'staff_portal_talent_profile',
  TalentCreate = 'staff_portal_talent_create',
  TalentUpdate = 'staff_portal_update_talent',
  TasksByClients = 'staff_portal_company_task_list',
  MyExpectedCommissions = 'my_expected_commissions',
  ClientProfile = 'staff_portal_company_profile',
  ClientsList = 'staff_portal_company_list',
  EntityHistory = 'staff_portal_entity_history',
  JobCreate = 'staff_portal_create_job',
  SourcedTalents = 'staff_portal_sourced_talents',
  Staff = 'staff_portal_managers',
  StaffProfile = 'staff_portal_staff_profile',
  TalentApplicants = 'staff_portal_talent_applicants',
  MeetingProfile = 'staff_portal_meeting_profile',
  SourcingRequestList = 'staff_portal_sourcing_request_list',
  TeamProfile = 'staff_portal_team_profile',
  EmailTemplates = 'staff_portal_email_templates'
}